move job-skills list into jobs page as ?view=skills
